feat: #5 apiで使用する型定義の記述をtypesへ移動
